refactor: helpers and tutor routes
