+++ src/utils/dataStructure/index.js
@@ -1,0 +1,8 @@
+const patientDataStructure = require('./patient.dataStructure');
+const activityDataStructure = require('./activity.dataStructure');
+
+
+module.exports = {
+  patientDataStructure,
+  activityDataStructure,
+}
